+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - Fix bug that causes materialization (V2) to fail when data type is long enough to be truncated by DESCRIBE TABLE ([1083](https://github.com/databricks/dbt-databricks/issues/1083))
 - Fix the bugs with external tabls cloning [1095](https://github.com/databricks/dbt-databricks/pull/1095) (thanks @samgans!)
 - Fix MV/ST materializations with complex data types ([1100](https://github.com/databricks/dbt-databricks/issues/1100))
+- fix: check constraints keep getting replaced [issue-1109](https://github.com/databricks/dbt-databricks/issues/1109)
 
 ### Under the Hood